# docs/swagger.yaml
basePath: /api/v1/
...
info:
  contact:
    email: ×××@qq.com
    name: API Support
    url: haimait.com
  description: Golang api of demo
  termsOfService: haimait.com
  title: haimait.com开发文档
  version: "1.0"
paths:
...
  /getUser/:id:
    get:
      consumes:
      - application/json
      description: 获取指定getUser记录2
      parameters:
      - description: 用户id
        in: query
        name: id
        required: true
        type: integer
      - description: 用户name
        in: query
        name: name
        type: string
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            type: string
      summary: 获取指定getUser记录1
      tags:
      - 测试
swagger: "2.0"